refactor dptb_negf as dpnegf
