+++ dpnegf/negf/bloch.py
