+++ dpnegf/negf/ozaki_res_cal.py
