+++ dpnegf/utils/constants.py
