+++ dpnegf/utils/tools.py
