+++ pyproject.toml
@@ -1,11 +1,11 @@
 [tool.poetry]
-name = "dptb-negf"
+name = "dpnegf"
 version = "0.1.0"
 license = "LGPL-3.0"
 description = "A nonequilibrium Green’s function (NEGF) module based on dptb."
 authors = ["J. Zou <jijiezou@stu.pku.edu.cn>","Q. Gu <guqq@pku.edu.cn>", "Z. Zhanghao <zhouyinzhanghao@gmail.com>"]
 readme = "README.md"
-repository = "https://github.com/DeePTB-Lab/DeePTB-negf"
+repository = "https://github.com/DeePTB-Lab/dpnegf"
 
 [tool.poetry.dependencies]
 python = ">=3.9, <=3.12.9"
@@ -70,7 +70,7 @@ build-backend = "poetry_dynamic_versioning.backend"
 [tool.poetry-dynamic-versioning]
 enable = true
 vcs = "git"
-strict = true
+strict = false
 format-jinja = """
     {%- if distance == 0 -%}
         {{ serialize_pep440(base, stage, revision) }}
